+++ CHANGELOG.rst
@@ -8,6 +8,7 @@ Unreleased
 Merged branches and closed issues
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
+* Feature: Added zhinst backend option to enable Calibration mode, Closes #103. (!123)
 * Added acquisitions to circuit diagram (!93)
 * Added Chirp and Staircase pulses; and efficient implementation in Qblox backend (!106)